+++ core/frontend/src/ViewportSync.ts
