+++ core/frontend/src/core-frontend.ts
@@ -58,7 +58,6 @@ export * from "./StandardView";
 export * from "./SubCategoriesCache";
 export * from "./TentativePoint";
 export * from "./Tiles";
-export * from "./TwoWayViewportSync";
 export * from "./UserPreferences";
 export * from "./ViewAnimation";
 export * from "./ViewContext";
@@ -66,6 +65,7 @@ export * from "./ViewGlobalLocation";
 export * from "./ViewingSpace";
 export * from "./ViewManager";
 export * from "./Viewport";
+export * from "./ViewportSync";
 export * from "./ViewPose";
 export * from "./ViewRect";
 export * from "./ViewState";
